refactor the copyright comments and DefaultMcpServerTest
